*: smattering of random small fixes to silence -Wall
I thought the build was already using -Wall but that seems to not
be the case, maybe got lost somewhere along the line or messed up
in configure.ac

After forcing a build with -Wall -Werror, these showed up.
Fixed up in the obvious way, nothing too scary.

+++ src/libs/din/din.c
@@ -18,7 +18,7 @@ typedef struct din_t {
 /* return random number between -1 and +1 */
 static inline float randf(unsigned *seed)
 {
-	return 2.f / RAND_MAX * rand_r(seed) - 1.f;
+	return 2.f / ((float)RAND_MAX) * rand_r(seed) - 1.f;
 }
